split io demod/pyramid demod to respective files
out of cmddata.c / data cmds
see changelog.md for cli changes as options have changed!
