+++ client/cmddata.c
@@ -943,233 +943,6 @@ int CmdFSKdemodParadox(const char *Cmd)
 	return 1;
 }
 
-//by marshmellow
-//IO-Prox demod - FSK RF/64 with preamble of 000000001
-//print ioprox ID and some format details
-int CmdFSKdemodIO(const char *Cmd)
-{
-	int idx=0;
-	//something in graphbuffer?
-	if (GraphTraceLen < 65) {
-		if (g_debugMode)PrintAndLog("DEBUG: not enough samples in GraphBuffer");
-		return 0;
-	}
-	uint8_t BitStream[MAX_GRAPH_TRACE_LEN]={0};
-	size_t BitLen = getFromGraphBuf(BitStream);
-	if (BitLen==0) return 0;
-
-	//get binary from fsk wave
-	idx = IOdemodFSK(BitStream,BitLen);
-	if (idx<0){
-		if (g_debugMode){
-			if (idx==-1){
-				PrintAndLog("DEBUG: Just Noise Detected");     
-			} else if (idx == -2) {
-				PrintAndLog("DEBUG: not enough samples");
-			} else if (idx == -3) {
-				PrintAndLog("DEBUG: error during fskdemod");        
-			} else if (idx == -4) {
-				PrintAndLog("DEBUG: Preamble not found");
-			} else if (idx == -5) {
-				PrintAndLog("DEBUG: Separator bits not found");
-			} else {
-				PrintAndLog("DEBUG: Error demoding fsk %d", idx);
-			}
-		}
-		return 0;
-	}
-	if (idx==0){
-		if (g_debugMode){
-			PrintAndLog("DEBUG: IO Prox Data not found - FSK Bits: %d",BitLen);
-			if (BitLen > 92) PrintAndLog("%s", sprint_bin_break(BitStream,92,16));
-		} 
-		return 0;
-	}
-		//Index map
-		//0           10          20          30          40          50          60
-		//|           |           |           |           |           |           |
-		//01234567 8 90123456 7 89012345 6 78901234 5 67890123 4 56789012 3 45678901 23
-		//-----------------------------------------------------------------------------
-		//00000000 0 11110000 1 facility 1 version* 1 code*one 1 code*two 1 ???????? 11
-		//
-		//XSF(version)facility:codeone+codetwo (raw)
-		//Handle the data
-	if (idx+64>BitLen) {
-		if (g_debugMode) PrintAndLog("not enough bits found - bitlen: %d",BitLen);
-		return 0;
-	}
-	PrintAndLog("%d%d%d%d%d%d%d%d %d",BitStream[idx],    BitStream[idx+1],  BitStream[idx+2], BitStream[idx+3], BitStream[idx+4], BitStream[idx+5], BitStream[idx+6], BitStream[idx+7], BitStream[idx+8]);
-	PrintAndLog("%d%d%d%d%d%d%d%d %d",BitStream[idx+9],  BitStream[idx+10], BitStream[idx+11],BitStream[idx+12],BitStream[idx+13],BitStream[idx+14],BitStream[idx+15],BitStream[idx+16],BitStream[idx+17]);
-	PrintAndLog("%d%d%d%d%d%d%d%d %d facility",BitStream[idx+18], BitStream[idx+19], BitStream[idx+20],BitStream[idx+21],BitStream[idx+22],BitStream[idx+23],BitStream[idx+24],BitStream[idx+25],BitStream[idx+26]);
-	PrintAndLog("%d%d%d%d%d%d%d%d %d version",BitStream[idx+27], BitStream[idx+28], BitStream[idx+29],BitStream[idx+30],BitStream[idx+31],BitStream[idx+32],BitStream[idx+33],BitStream[idx+34],BitStream[idx+35]);
-	PrintAndLog("%d%d%d%d%d%d%d%d %d code1",BitStream[idx+36], BitStream[idx+37], BitStream[idx+38],BitStream[idx+39],BitStream[idx+40],BitStream[idx+41],BitStream[idx+42],BitStream[idx+43],BitStream[idx+44]);
-	PrintAndLog("%d%d%d%d%d%d%d%d %d code2",BitStream[idx+45], BitStream[idx+46], BitStream[idx+47],BitStream[idx+48],BitStream[idx+49],BitStream[idx+50],BitStream[idx+51],BitStream[idx+52],BitStream[idx+53]);
-	PrintAndLog("%d%d%d%d%d%d%d%d %d%d checksum",BitStream[idx+54],BitStream[idx+55],BitStream[idx+56],BitStream[idx+57],BitStream[idx+58],BitStream[idx+59],BitStream[idx+60],BitStream[idx+61],BitStream[idx+62],BitStream[idx+63]);
-
-	uint32_t code = bytebits_to_byte(BitStream+idx,32);
-	uint32_t code2 = bytebits_to_byte(BitStream+idx+32,32);
-	uint8_t version = bytebits_to_byte(BitStream+idx+27,8); //14,4
-	uint8_t facilitycode = bytebits_to_byte(BitStream+idx+18,8) ;
-	uint16_t number = (bytebits_to_byte(BitStream+idx+36,8)<<8)|(bytebits_to_byte(BitStream+idx+45,8)); //36,9
-	uint8_t crc = bytebits_to_byte(BitStream+idx+54,8);
-	uint16_t calccrc = 0;
-
-	for (uint8_t i=1; i<6; ++i){
-		calccrc += bytebits_to_byte(BitStream+idx+9*i,8);
-	}
-	calccrc &= 0xff;
-	calccrc = 0xff - calccrc;
-
-	char *crcStr = (crc == calccrc) ? "crc ok": "!crc";
-
-	PrintAndLog("IO Prox XSF(%02d)%02x:%05d (%08x%08x) [%02x %s]",version,facilitycode,number,code,code2, crc, crcStr);
-	setDemodBuf(BitStream,64,idx);
-	if (g_debugMode){
-		PrintAndLog("DEBUG: idx: %d, Len: %d, Printing demod buffer:",idx,64);
-		printDemodBuff();
-	}
-	return 1;
-}
-
-//by marshmellow
-//Pyramid Prox demod - FSK RF/50 with preamble of 0000000000000001  (always a 128 bit data stream)
-//print full Farpointe Data/Pyramid Prox ID and some bit format details if found
-int CmdFSKdemodPyramid(const char *Cmd)
-{
-	//raw fsk demod no manchester decoding no start bit finding just get binary from wave
-	uint8_t BitStream[MAX_GRAPH_TRACE_LEN]={0};
-	size_t size = getFromGraphBuf(BitStream);
-	if (size==0) return 0;
-
-	//get binary from fsk wave
-	int idx = PyramiddemodFSK(BitStream, &size);
-	if (idx < 0){
-		if (g_debugMode){
-			if (idx == -5)
-				PrintAndLog("DEBUG: Error - not enough samples");
-			else if (idx == -1)
-				PrintAndLog("DEBUG: Error - only noise found");
-			else if (idx == -2)
-				PrintAndLog("DEBUG: Error - problem during FSK demod");
-			else if (idx == -3)
-				PrintAndLog("DEBUG: Error - Size not correct: %d", size);
-			else if (idx == -4)
-				PrintAndLog("DEBUG: Error - Pyramid preamble not found");
-			else
-				PrintAndLog("DEBUG: Error - idx: %d",idx);
-		}
-		return 0;
-	}
-	// Index map
-	// 0           10          20          30            40          50          60
-	// |           |           |           |             |           |           |
-	// 0123456 7 8901234 5 6789012 3 4567890 1 2345678 9 0123456 7 8901234 5 6789012 3
-	// -----------------------------------------------------------------------------
-	// 0000000 0 0000000 1 0000000 1 0000000 1 0000000 1 0000000 1 0000000 1 0000000 1
-	// premable  xxxxxxx o xxxxxxx o xxxxxxx o xxxxxxx o xxxxxxx o xxxxxxx o xxxxxxx o
-
-	// 64    70            80          90          100         110           120
-	// |     |             |           |           |           |             |
-	// 4567890 1 2345678 9 0123456 7 8901234 5 6789012 3 4567890 1 2345678 9 0123456 7
-	// -----------------------------------------------------------------------------
-	// 0000000 1 0000000 1 0000000 1 0110111 0 0011000 1 0000001 0 0001100 1 1001010 0
-	// xxxxxxx o xxxxxxx o xxxxxxx o xswffff o ffffccc o ccccccc o ccccccw o ppppppp o
-	//                                  |---115---||---------71---------|
-	// s = format start bit, o = odd parity of last 7 bits
-	// f = facility code, c = card number
-	// w = wiegand parity, x = extra space for other formats
-	// p = unknown checksum
-	// (26 bit format shown)
-
-	//get bytes for checksum calc
-	uint8_t checksum = bytebits_to_byte(BitStream + idx + 120, 8);
-	uint8_t csBuff[14] = {0x00};
-	for (uint8_t i = 0; i < 13; i++){
-		csBuff[i] = bytebits_to_byte(BitStream + idx + 16 + (i*8), 8);
-	}
-	//check checksum calc
-	//checksum calc thanks to ICEMAN!!
-	uint32_t checkCS =  CRC8Maxim(csBuff,13);
-
-	//get raw ID before removing parities
-	uint32_t rawLo = bytebits_to_byte(BitStream+idx+96,32);
-	uint32_t rawHi = bytebits_to_byte(BitStream+idx+64,32);
-	uint32_t rawHi2 = bytebits_to_byte(BitStream+idx+32,32);
-	uint32_t rawHi3 = bytebits_to_byte(BitStream+idx,32);
-	setDemodBuf(BitStream,128,idx);
-
-	size = removeParity(BitStream, idx+8, 8, 1, 120);
-	if (size != 105){
-		if (g_debugMode) 
-			PrintAndLog("DEBUG: Error at parity check - tag size does not match Pyramid format, SIZE: %d, IDX: %d, hi3: %x",size, idx, rawHi3);
-		return 0;
-	}
-
-	// ok valid card found!
-
-	// Index map
-	// 0         10        20        30        40        50        60        70
-	// |         |         |         |         |         |         |         |
-	// 01234567890123456789012345678901234567890123456789012345678901234567890
-	// -----------------------------------------------------------------------
-	// 00000000000000000000000000000000000000000000000000000000000000000000000
-	// xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-
-	// 71         80         90          100
-	// |          |          |           |
-	// 1 2 34567890 1234567890123456 7 8901234
-	// ---------------------------------------
-	// 1 1 01110011 0000000001000110 0 1001010
-	// s w ffffffff cccccccccccccccc w ppppppp
-	//     |--115-| |------71------|
-	// s = format start bit, o = odd parity of last 7 bits
-	// f = facility code, c = card number
-	// w = wiegand parity, x = extra space for other formats
-	// p = unknown checksum
-	// (26 bit format shown)
-
-	//find start bit to get fmtLen
-	int j;
-	for (j=0; j<size; j++){
-		if(BitStream[j]) break;
-	}
-	uint8_t fmtLen = size-j-8;
-	uint32_t fc = 0;
-	uint32_t cardnum = 0;
-	uint32_t code1 = 0;
-	if (fmtLen==26){
-		fc = bytebits_to_byte(BitStream+73, 8);
-		cardnum = bytebits_to_byte(BitStream+81, 16);
-		code1 = bytebits_to_byte(BitStream+72,fmtLen);
-		PrintAndLog("Pyramid ID Found - BitLength: %d, FC: %d, Card: %d - Wiegand: %x, Raw: %08x%08x%08x%08x", fmtLen, fc, cardnum, code1, rawHi3, rawHi2, rawHi, rawLo);
-	} else if (fmtLen==45){
-		fmtLen=42; //end = 10 bits not 7 like 26 bit fmt
-		fc = bytebits_to_byte(BitStream+53, 10);
-		cardnum = bytebits_to_byte(BitStream+63, 32);
-		PrintAndLog("Pyramid ID Found - BitLength: %d, FC: %d, Card: %d - Raw: %08x%08x%08x%08x", fmtLen, fc, cardnum, rawHi3, rawHi2, rawHi, rawLo);
-	} else {
-		cardnum = bytebits_to_byte(BitStream+81, 16);
-		if (fmtLen>32){
-			//code1 = bytebits_to_byte(BitStream+(size-fmtLen),fmtLen-32);
-			//code2 = bytebits_to_byte(BitStream+(size-32),32);
-			PrintAndLog("Pyramid ID Found - BitLength: %d -unknown BitLength- (%d), Raw: %08x%08x%08x%08x", fmtLen, cardnum, rawHi3, rawHi2, rawHi, rawLo);
-		} else{
-			//code1 = bytebits_to_byte(BitStream+(size-fmtLen),fmtLen);
-			PrintAndLog("Pyramid ID Found - BitLength: %d -unknown BitLength- (%d), Raw: %08x%08x%08x%08x", fmtLen, cardnum, rawHi3, rawHi2, rawHi, rawLo);
-		}
-	}
-	if (checksum == checkCS)
-		PrintAndLog("Checksum %02x passed", checksum);
-	else
-		PrintAndLog("Checksum %02x failed - should have been %02x", checksum, checkCS);
-
-	if (g_debugMode){
-		PrintAndLog("DEBUG: idx: %d, Len: %d, Printing Demod Buffer:", idx, 128);
-		printDemodBuff();
-	}
-	return 1;
-}
-
 // FDX-B ISO11784/85 demod  (aka animal tag)  BIPHASE, inverted, rf/32,  with preamble of 00000000001 (128bits)
 // 8 databits + 1 parity (1)
 // CIITT 16 chksum
@@ -2094,8 +1867,6 @@ static command_t CommandTable[] =
 	{"detectclock",     CmdDetectClockRate, 1, "[modulation] Detect clock rate of wave in GraphBuffer (options: 'a','f','n','p' for ask, fsk, nrz, psk respectively)"},
 	{"fdxbdemod",       CmdFDXBdemodBI    , 1, "Demodulate a FDX-B ISO11784/85 Biphase tag from GraphBuffer"},
 	//{"fskfcdetect",   CmdFSKfcDetect,     1, "Try to detect the Field Clock of an FSK wave"},
-	{"fskiodemod",      CmdFSKdemodIO,      1, "Demodulate an IO Prox FSK tag from GraphBuffer"},
-	{"fskpyramiddemod", CmdFSKdemodPyramid, 1, "Demodulate a Pyramid FSK tag from GraphBuffer"},
 	{"fskparadoxdemod", CmdFSKdemodParadox, 1, "Demodulate a Paradox FSK tag from GraphBuffer"},
 	{"getbitstream",    CmdGetBitStream,    1, "Convert GraphBuffer's >=1 values to 1 and <1 to 0"},
 	{"grid",            CmdGrid,            1, "<x> <y> -- overlay grid on graph window, use zero value to turn off either"},
